setup: default cmd and i18n config
