+++ README.md
@@ -1,5 +1,20 @@
 # Pythonlings
 
+Pythonlings is a multilingual utility to help newcomers learn Python through exercises, with a variety of themes!
+
+## Setup your language
+
+Actually, the following languages are supported:
+    * en (default)
+    * pt
+
+To choose your language, just export `PYTHONLINGS_LANGUAGE` with the desired language.
+Example:
+
+
+    export PYTHONLINGS_LANGUAGE=pt
+
+
 ## Usage
 
     poetry run -m pythonlings
