+++ pythonlings/services/logging.py
@@ -1,0 +1,6 @@
+import logging
+
+
+logging.basicConfig(level=logging.INFO,
+    format='%(asctime)s - %(message)s',
+    datefmt='%Y-%m-%d %H:%M:%S')
